+++ docs/30_narrative/character-profiles/hero.md
