+++ docs/30_narrative/character-profiles/npc.md
