+++ docs/30_narrative/story-outline.md
