+++ docs/30_narrative/world-setting.md
