+++ docs/40_technical/input-spec.md
