+++ public/src/core/asset-loader.js
