+++ public/src/core/scene-manager.js
