+++ public/src/entities/npc.js
